requirements.txt
async-timeout==4.0.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
fonttools==4.38.0
gensim==4.2.0
hdbscan==0.8.29
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.6.2
mypy-extensions==0.4.3
numba==0.56.4
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.5.4
pynndescent==0.5.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
redis==4.3.5
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
smart-open==6.2.0
spotipy==2.21.0
threadpoolctl==3.1.0
tomli==2.0.1
top2vec==1.0.27
tqdm==4.64.1
umap-learn==0.5.3
urllib3==1.26.13
wordcloud==1.8.2.2
